keep rules sort order consistent after update
The first Drag-n-Drop through Sortablejs would generate a correct
update call to the API but the data in the Vue client would not
be updated correctly. It seems a copy of the HTML element is kept
in the Sortable component which is not the same as the one Vue is
working with.